# swagger/swagger.yaml
basePath: /rest/api/1
consumes:
- application/json
definitions:
  APIKey:
    properties:
      api_key_id:
        description: APIKeyID is a generated id of the key.
        type: string
        x-go-name: APIKeyID
      api_key_secret:
        description: APIKeySecret is a generated secret, aka, the key.
        type: string
        x-go-name: APIKeySecret
      create_at:
        description: CreateAt defines when this key was created.
        example: time.Now()
        format: date-time
        type: string
        x-go-name: CreateAt
      id:
        description: ID of the key. This is auto-generated.
        format: int64
        type: integer
        x-go-name: ID
      name:
        description: Name of the key
        type: string
        x-go-name: Name
      ttl:
        description: TTL defines how long this key can live in duration.
        example: 1h10m10s
        type: string
        x-go-name: TTL
      user_id:
        description: UserID is the ID of the user to which this key belongs.
        format: int64
        type: integer
        x-go-name: UserID
    required:
    - id
    - name
    - user_id
    - api_key_id
    - api_key_secret
    - ttl
    - create_at
    title: APIKey is an api key pair generated by the user to access the api without
      the frontend.
    type: object
    x-go-package: github.com/krok-o/krok/pkg/models
...
host: krok-o.io
info:
  description: Documentation the Krok API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Krok.
  version: 0.0.1
paths:
...
  /user/apikey:
    post:
      operationId: listApiKeys
      produces:
      - application/json
      responses:
        "200":
          description: ""
          schema:
            items:
              $ref: '#/definitions/APIKey'
            type: array
        "500":
          description: failed to get user context
          schema:
            $ref: '#/responses/Message'
      summary: Lists all api keys for a given user.
  /user/apikey/{keyid}:
    get:
      operationId: getApiKeys
      parameters:
      - description: The ID of the key to return
        format: int
        in: path
        name: keyid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
          schema:
            $ref: '#/definitions/APIKey'
        "500":
          description: failed to get user context
          schema:
            $ref: '#/responses/Message'
      summary: Returns a given api key.
  /user/apikey/delete/{keyid}:
    delete:
      operationId: deleteApiKey
      parameters:
      - description: The ID of the key to delete
        format: int
        in: path
        name: keyid
        required: true
        type: integer
      responses:
        "200":
          description: OK in case the deletion was successful
        "400":
          description: in case of missing user context or invalid ID
          schema:
            $ref: '#/responses/Message'
        "500":
          description: when the deletion operation failed
          schema:
            $ref: '#/responses/Message'
      summary: Deletes a set of api keys for a given user with a given id.
  /user/apikey/generate/{name}:
    post:
      operationId: createApiKey
      parameters:
      - description: the name of the key
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: the generated api key pair
          schema:
            $ref: '#/definitions/APIKey'
        "400":
          description: failed to generate unique key or value
          schema:
            $ref: '#/responses/Message'
        "500":
          description: when failed to get user context
          schema:
            $ref: '#/responses/Message'
      summary: Creates an api key pair for a given user.
...
produces:
- application/json
responses:
  Message:
    description: Message represents an error message.
...
schemes:
- http
swagger: "2.0"